requirements.txt
discord.py==1.7.3
projects==1.0.9
psycopg2==2.9.3
psycopg2_binary==2.8.6
python-dotenv==0.20.0
